requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
flask-restx==1.3.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
urllib3==2.1.0
Werkzeug==3.0.1
